requirements.txt
streamlit
google-generativeai
PyMuPDF
# trigger redeploy
